SImplify settings UI
